requirements.txt
flask>=0.12.2
gunicorn>=19.7.0
gevent>=1.4
weaviate-client>=3.23.0
python-dotenv>=1.0.0